added comments

+++ packages/protocol/contracts/identity/test/TestRandom.sol
@@ -6,7 +6,7 @@ contract TestRandom is Random {
   function addTestRandomness(uint256 blockNumber, bytes32 randomness) external {
     addRandomness(blockNumber, randomness);
   }
-  function getTestRandomness(uint blockNumber, uint cur) external view returns (bytes32) {
+  function getTestRandomness(uint256 blockNumber, uint256 cur) external view returns (bytes32) {
     return _getBlockRandomness(blockNumber, cur);
   }
 }
